Refactor NGS file reader, include skipping lines with comments
